separating concerns in db
