+++ package.json
@@ -7,7 +7,7 @@
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "nodemon ./server/index.js",
     "react-dev": "webpack --watch",
-    "seed-db": "node ./database/schema"
+    "seed-db": "node ./database/seed_db.js"
   },
   "repository": {
     "type": "git",
@@ -24,6 +24,7 @@
     "babel": "^6.23.0",
     "babel-loader": "^8.0.6",
     "express": "^4.17.1",
+    "faker": "^4.1.0",
     "live-server": "^1.2.1",
     "mongo": "^0.1.0",
     "mongoose": "^5.7.4",
